fix don't delete user if it is in use

+++ task_manager/users/views.py
@@ -5,6 +5,8 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.utils.translation import gettext_lazy as _
 from django.views import View
 
+from task_manager.tasks.models import Tasks
+
 from .forms import CreateUserForm
 from .models import Users
 
@@ -115,6 +117,9 @@ class DeleteUserView(BaseUserView):
         user = self._get_user(pk)
         if isinstance(user, HttpResponseRedirect):
             return user
+        if Tasks.objects.filter(executor=user).exists():
+            messages.error(request, _('Cannot delete user because it is in use'))
+            return redirect('users')
         user.delete()
         messages.success(request, _('User successfully deleted'))
         return redirect('users')
